deploy/pyproject.toml
[project]
name = "christmas-namedraw"
version = "0.2.0"
description = "Elf-themed name draw experience powered by Flask and Microsoft Graph"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Buddy the Elf" }]
dependencies = [
    "flask>=3.0",
    "requests>=2.31",
    "pyjwt[crypto]>=2.8",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/rtreit/christmas-namedraw"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4",
    "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
include = ["namedraw"]
